requirements.txt
flask
flask-cors
rembg
onnxruntime
Pillow